+++ android/src/main/java/com/getcapacitor/community/media/photoviewer/helper/CallbackListener.kt
@@ -1,0 +1,5 @@
+package com.getcapacitor.community.media.photoviewer.helper
+
+interface CallbackListener {
+    fun onMenuToggle()
+}