+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.0.1-beta.1 (2021-03-08) 
+
+## Chore
+
+ - Update to @capacitor/android@3.0.0-beta.6
+ 
+## Added Features
+
+ - Android Rotation Portrait to Landscape 
+ - Android Close button when image in fullscreen
+ - Android Hiding Share and Close button on singleTap
+
 # 0.0.1-alpha.2 (2021-03-06) 
 
 ## Added Features